Write config.xml for ios on update
